extracting the queued notification events persistence to its own file
