+++ vector/src/main/java/im/vector/app/features/notifications/NotificationState.kt
@@ -18,13 +18,6 @@ package im.vector.app.features.notifications
 
 import android.content.Context
 import org.matrix.android.sdk.api.session.Session
-import timber.log.Timber
-import java.io.File
-import java.io.FileOutputStream
-
-// TODO Mutliaccount
-private const val ROOMS_NOTIFICATIONS_FILE_NAME = "im.vector.notifications.cache"
-private const val KEY_ALIAS_SECRET_STORAGE = "notificationMgr"
 
 class NotificationState(
         /**
@@ -65,57 +58,20 @@ class NotificationState(
 
     fun hasAlreadyRendered(eventsToRender: List<ProcessedEvent<NotifiableEvent>>) = renderedEvents == eventsToRender
 
-    fun persistState(context: Context, currentSession: Session) {
-        synchronized(queuedEvents) {
-            if (queuedEvents.isEmpty()) {
-                deleteCachedRoomNotifications(context)
-                return
-            }
-            try {
-                val file = File(context.applicationContext.cacheDir, ROOMS_NOTIFICATIONS_FILE_NAME)
-                if (!file.exists()) file.createNewFile()
-                FileOutputStream(file).use {
-                    currentSession.securelyStoreObject(queuedEvents.rawEvents(), KEY_ALIAS_SECRET_STORAGE, it)
-                }
-            } catch (e: Throwable) {
-                Timber.e(e, "## Failed to save cached notification info")
-            }
-        }
-    }
-
-    private fun deleteCachedRoomNotifications(context: Context) {
-        val file = File(context.applicationContext.cacheDir, ROOMS_NOTIFICATIONS_FILE_NAME)
-        if (file.exists()) {
-            file.delete()
-        }
+    fun persist(context: Context, session: Session) {
+        NotificationEventPersistence.persistEvents(queuedEvents, context, session)
     }
 
     companion object {
 
         fun createInitialNotificationState(context: Context, currentSession: Session?): NotificationState {
-            val queuedEvents = loadEventInfo(context, currentSession)
+            val queuedEvents = NotificationEventPersistence.loadEvents(context, currentSession)
             return NotificationState(
                     queuedEvents = queuedEvents,
                     renderedEvents = queuedEvents.rawEvents().map { ProcessedEvent(ProcessedEvent.Type.KEEP, it) }.toMutableList(),
                     seenEventIds = CircularCache.create(cacheSize = 25)
             )
         }
-
-        private fun loadEventInfo(context: Context, currentSession: Session?): NotificationEventQueue {
-            try {
-                val file = File(context.applicationContext.cacheDir, ROOMS_NOTIFICATIONS_FILE_NAME)
-                if (file.exists()) {
-                    file.inputStream().use {
-                        val events: ArrayList<NotifiableEvent>? = currentSession?.loadSecureSecret(it, KEY_ALIAS_SECRET_STORAGE)
-                        if (events != null) {
-                            return NotificationEventQueue(events.toMutableList())
-                        }
-                    }
-                }
-            } catch (e: Throwable) {
-                Timber.e(e, "## Failed to load cached notification info")
-            }
-            return NotificationEventQueue()
-        }
     }
 }
+
